src/requirements.txt
Flask
pymysql
flask-mysqldb
python-dotenv
python-decouple
pydantic[email]
email-validator
cryptography
